requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
attrs==16.2.0
dj-database-url==0.4.1
django-graphiql==0.4.4
Django==1.10.1            # via django-graphiql, graphene-django, graphql-django-view
graphene-django==1.0.dev20160910000001
graphene==1.0.dev20160911051803
graphql-core==1.0.dev20160909040033  # via graphene, graphql-django-view, graphql-relay
graphql-django-view==1.3  # via graphene-django
graphql-relay==0.4.4      # via graphene
gunicorn==19.6.0
iso8601==0.1.11           # via graphene-django
promise==0.4.2            # via graphene, graphql-core, graphql-relay
psycopg2==2.6.2
singledispatch==3.4.0.3   # via graphene-django
six==1.10.0               # via graphene, graphene-django, graphql-core, singledispatch
